+++ exports/components/light-preview/light-preview.js
@@ -84,6 +84,7 @@ export default class LightPreview extends BaseElement {
   static properties = {
     summary: {},
     sandboxSettings: { reflect: true, attribute: "sandbox-settings" },
+    src: {},
     previewMode: { reflect: true, attribute: "preview-mode" },
     open: { reflect: true, type: Boolean },
     resizePosition: {
@@ -118,10 +119,16 @@ export default class LightPreview extends BaseElement {
     super();
 
     /**
+     * The `iframe` src to use. This can be set to an external URL. This will override any HTML provided to the component.
+     */
+    this.src = "";
+
+    /**
      * The sandbox settings to provide to the <iframe>
      * @type {string}
      */
-    this.sandboxSettings = "";
+    this.sandboxSettings =
+      "allow-downloads allow-forms allow-modals allow-pointer-lock allow-popups allow-presentation allow-scripts";
 
     /**
      * The text to provide in the <details> toggle button
@@ -407,7 +414,27 @@ export default class LightPreview extends BaseElement {
         </body>
       </html>`;
 
-    iframe.srcdoc = content;
+    if (!this.src) {
+      iframe.srcdoc = content;
+    } else {
+      iframe.removeAttribute("srcdoc");
+    }
+  }
+
+  /**
+   * @override
+   * @param {import("lit").PropertyValues<typeof this>} changedProperties
+   */
+  updated(changedProperties) {
+    const iframe = this.shadowRoot?.querySelector("iframe");
+    if (iframe && this.src) {
+      iframe.removeAttribute("srcdoc");
+    }
+
+    if (changedProperties.has("src") && !this.src) {
+      this.updateIframeContent();
+    }
+    super.updated(changedProperties);
   }
 
   /**
@@ -574,6 +601,7 @@ export default class LightPreview extends BaseElement {
    * @override
    */
   render() {
+    const shouldUseSrcdoc = !this.src;
     const finalHTML = html`
       <div
         part=${stringMap({
@@ -589,6 +617,7 @@ export default class LightPreview extends BaseElement {
                 part="start-panel iframe"
                 height="auto"
                 frameborder="0"
+                src=${this.src}
                 sandbox=${ifDefined(
                   this.sandboxSettings ? this.sandboxSettings : null,
                 )}