+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 4.3.0 - 07/04/2025
 
+- Security Fix(light-pen / light-preview): Previously, `<iframe>` in these 2 components could reach the parent. With this fix, it now has a default set of sandbox settings as follows: `allow-downloads allow-forms allow-modals allow-pointer-lock allow-popups allow-presentation allow-scripts`
 - Bug Fix(light-editor): Fixed a bug with grid setup for soft wrapping
+- Feature(light-preview): Added support for `src` attribute for external URLs.
 
 ## 4.2.6 - 05/02/2025